future-app/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
anthropic>=0.50.0
python-dotenv==1.0.1
alembic>=1.13.1
Mako>=1.3.0
yfinance>=0.2.40
requests>=2.31.0
